Resetting indices for joint_df
There is a bug when using loading a separate testing dataframe when adding the Energy based weights, it duplicates the structure due to having the same indices repeated. 

For example, 

My training dataframe had 4500 configurations while the testing dataframe had 4111 configurations, combined they are 8611 configurations. The saved datasets with weights would have duplicate entries and it would be nearly double the size of the dataframe.

The suggested fix is to simply reset the indices for the joint dataframe before applying the weights.